rsvp: Add beginnings of a CHICKEN Scheme RSVP backend
The plan is to host this at rsvp.jennex.org (separately from the
static site) and have it act as a dynamic server for the RSVP form.

It will have to:

* Fetch and display existing RSVP's
* Group guests
* Allow updating RSVP's

The plan is to do this all in Scheme for the fun and the learnings.
It's already proven educational getting my dev environment set up :)