rsvp: Update success page edit link to go to form
There's a bit of a hack here where if the second person in the party
hits the edit link, it might have the first person's name, but hopefully
people won't mind / notice.

Also this breaks the "maybe we should make the edit page a POST request
so that it's not linkable" TODO... but that's an issue for later.