rsvp: Overhaul party and plus1 system (now with notes!)
I hadn't implemented any of the "plus1" for individuals, and I think a
cleaner route will just be to plug each plus1 in as a "Guest" for the
party with the name "Foo Bar's Plus 1" and then just allow the party
members to update their names.

The main meat of this change was updating the code to handle a party
instead of individuals. Fetching a whole party is a nice way to handle
the form.

Finally, this patch updates the form to have a textarea for party-notes,
which will be important for misc. things guests need us to know!